requirements.txt
PyYAML
shell
requests
